+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/FieldScope.java
@@ -37,7 +37,7 @@ public class FieldScope extends MemberScope<ResolvedField, Field> {
      * @param context the overall type resolution context
      */
     protected FieldScope(ResolvedField field, ResolvedTypeWithMembers declaringTypeMembers, TypeContext context) {
-        this(field, null, null, declaringTypeMembers, context);
+        this(field, null, null, declaringTypeMembers, false, context);
     }
 
     /**
@@ -47,21 +47,24 @@ public class FieldScope extends MemberScope<ResolvedField, Field> {
      * @param overriddenType alternative type for this field
      * @param overriddenName alternative name for this field
      * @param declaringTypeMembers collection of the declaring type's (other) fields and methods
+     * @param fakeContainerItemScope whether this field/method scope represents only the container item type of the actual field/method
      * @param context the overall type resolution context
      */
     protected FieldScope(ResolvedField field, ResolvedType overriddenType, String overriddenName,
-            ResolvedTypeWithMembers declaringTypeMembers, TypeContext context) {
-        super(field, overriddenType, overriddenName, declaringTypeMembers, context);
+            ResolvedTypeWithMembers declaringTypeMembers, boolean fakeContainerItemScope, TypeContext context) {
+        super(field, overriddenType, overriddenName, declaringTypeMembers, fakeContainerItemScope, context);
     }
 
     @Override
     public FieldScope withOverriddenType(ResolvedType overriddenType) {
-        return new FieldScope(this.getMember(), overriddenType, this.getOverriddenName(), this.getDeclaringTypeMembers(), this.getContext());
+        return new FieldScope(this.getMember(), overriddenType, this.getOverriddenName(), this.getDeclaringTypeMembers(),
+                this.isFakeContainerItemScope(), this.getContext());
     }
 
     @Override
     public FieldScope withOverriddenName(String overriddenName) {
-        return new FieldScope(this.getMember(), this.getOverriddenType(), overriddenName, this.getDeclaringTypeMembers(), this.getContext());
+        return new FieldScope(this.getMember(), this.getOverriddenType(), overriddenName, this.getDeclaringTypeMembers(),
+                this.isFakeContainerItemScope(), this.getContext());
     }
 
     /**